ewm-main-svc Продорлжаем бороться с ошибками

+++ ewm-main-svc/src/main/java/ru/practicum/ewm/event/model/Event.java
@@ -63,6 +63,6 @@ public class Event {
 
     @Column(name = "confirmed_requests")
     int confirmedRequests;
-    int views;
+    long views;
 }
